% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolRollNZ.R
\docType{data}
\name{school_deciles}
\alias{school_deciles}
\title{School Deciles, multiple entries if decile changed between 2000 and 2017}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{School ID}{}
\item{School Name}{}
\item{Decile}{}
\item{Years}{}
\item{Number of deciles}{}
}}
\source{
The data comes from https://www.educationcounts.govt.nz/statistics/schooling/student-numbers/6028
}
\description{
Detailed description of the data
}
\seealso{
\link{SchoolRollNZ}
\link{school_demographics}
\link{school_details}
}
