% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolRollNZ.R
\docType{data}
\name{school_demographics}
\alias{school_demographics}
\title{Longform table with school rolls and (some) demographic details}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{School ID}{}
\item{School Name}{}
\item{Year}{}
\item{Demographic}{}
\item{Count}{}
}}
\source{
The data comes from https://www.educationcounts.govt.nz/statistics/schooling/student-numbers/6028
}
\description{
Detailed description of the data
}
\seealso{
\link{SchoolRollNZ}
\link{school_deciles}
\link{school_details}
}
