% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolRollNZ.R
\docType{data}
\name{school_details}
\alias{school_details}
\title{School details, changes in detail between 2000 and 2017 result in multiple entries}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{School ID}{}
\item{School Name}{}
\item{Type}{}
\item{Authority}{}
\item{Gender}{}
\item{Regional Council}{}
\item{TA with Auckland Local Board}{}
\item{Years}{}
\item{Number of changes}{}
}}
\source{
The data comes from https://www.educationcounts.govt.nz/statistics/schooling/student-numbers/6028
}
\description{
Detailed description of the data
}
\seealso{
\link{SchoolRollNZ}
\link{school_deciles}
\link{school_demographics}
}
