% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolRollNZ.R
\docType{package}
\name{SchoolRollNZ}
\alias{SchoolRollNZ}
\alias{SchoolRollNZ-package}
\title{New Zealand School Rolls}
\description{
(Some of) the school roll data released by the Ministry of Education in longform.
}
\details{
SchoolRollNZ
A data package of New Zealand School Rolls.

Use \code{data(package='SchoolRollNZ')$results[, 3]} tosee a list of availabledata sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{school_deciles}
\link{school_demographics}
\link{school_details}
}
